rename SeqAutomation.h to SeqMotionLanes.h
